[Q20-3] - addressed review comments - moved authentication files to appropriate packages
